+++ packages/amplify-graphql-api-construct-tests/src/__tests__/conversations/graphql/subscriptions.ts
@@ -56,6 +56,7 @@ export const onCreateAssistantResponsePirateChat = /* GraphQL */ `subscription O
     id
     owner
     stopReason
+    p
   }
 }
 ` as GeneratedSubscription<