+++ packages/amplify-graphql-conversation-transformer/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.7](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-conversation-transformer@1.1.6...@aws-amplify/graphql-conversation-transformer@1.1.7) (2025-02-07)
+
+### Bug Fixes
+
+- bump ai-constructs to ^1.2.2 ([#3141](https://github.com/aws-amplify/amplify-category-api/issues/3141)) ([23519e3](https://github.com/aws-amplify/amplify-category-api/commit/23519e3a682175ea6e02c41b5fecac5c51ce04bf))
+
+## [1.1.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-conversation-transformer@1.1.5...@aws-amplify/graphql-conversation-transformer@1.1.6) (2025-01-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-conversation-transformer
+
+## [1.1.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-conversation-transformer@1.1.3...@aws-amplify/graphql-conversation-transformer@1.1.5) (2025-01-16)
+
+### Bug Fixes
+
+- **conversation:** add padding to stream chunks ([#3115](https://github.com/aws-amplify/amplify-category-api/issues/3115)) ([8ab2284](https://github.com/aws-amplify/amplify-category-api/commit/8ab2284eedc7f5e29d0e21acf3124d60ed058a93))
+
 ## [1.1.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-conversation-transformer@1.1.3...@aws-amplify/graphql-conversation-transformer@1.1.4) (2024-12-23)
 
 **Note:** Version bump only for package @aws-amplify/graphql-conversation-transformer