+++ packages/amplify-graphql-transformer-interfaces/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@4.2.2...@aws-amplify/graphql-transformer-interfaces@4.2.3) (2025-02-07)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces
+
+## [4.2.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@4.2.1...@aws-amplify/graphql-transformer-interfaces@4.2.2) (2025-01-30)
+
+### Bug Fixes
+
+- **gen2:** transform directives on fields of supported extended types ([#3127](https://github.com/aws-amplify/amplify-category-api/issues/3127)) ([3aca7b8](https://github.com/aws-amplify/amplify-category-api/commit/3aca7b89f530c536c28a53b7290fdab877b2260b))
+
 ## [4.2.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-interfaces@4.2.0...@aws-amplify/graphql-transformer-interfaces@4.2.1) (2024-12-17)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-interfaces